[wip] map measure refactore from PR #171
